// ui/src/api/generated/models/model-parameter-mappings.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.25.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EmbeddingMappings } from './embedding-mappings';
// May contain unused imports in some cases
// @ts-ignore
import type { ImageGenerationMappings } from './image-generation-mappings';
// May contain unused imports in some cases
// @ts-ignore
import type { LanguageMappings } from './language-mappings';
// May contain unused imports in some cases
// @ts-ignore
import type { RerankMappings } from './rerank-mappings';

/**
 * Administrator overrides for provider-neutral parameter mappings
 * @export
 * @interface ModelParameterMappings
 */
export interface ModelParameterMappings {
    /**
     * 
     * @type {EmbeddingMappings}
     * @memberof ModelParameterMappings
     */
    'embedding'?: EmbeddingMappings;
    /**
     * 
     * @type {ImageGenerationMappings}
     * @memberof ModelParameterMappings
     */
    'imageGeneration'?: ImageGenerationMappings;
    /**
     * 
     * @type {LanguageMappings}
     * @memberof ModelParameterMappings
     */
    'language'?: LanguageMappings;
    /**
     * 
     * @type {RerankMappings}
     * @memberof ModelParameterMappings
     */
    'rerank'?: RerankMappings;
}